Rework app model
This is a major rework of the app’s model. It changes how the update flow throughout the app is orchestrated. App objects are now immutable and split up in multiple objects (Bundle and Update) which better represent their semantical purposes. This should avoid data races which ultimately result in crashes like #114.

.