Open applications asynchronously in macOS 10.15 and later
Latest opens applications with the [synchronous method](https://developer.apple.com/documentation/appkit/nsworkspace/1533463-open), which makes it wait for the opening. This can be a long time for big apps (or if online connectivity is spotty), during which Latest remains unresponsive.

To fix that, we can use the [asynchronous method](https://developer.apple.com/documentation/appkit/nsworkspace/3172701-open), which doesn't make the thread (and the app) wait. It is only available in macOS 10.15+, so we wrap it in an `@available` check.

The same way that in the previous code we ignored the return value, we can ignore the `completionHandler` here too.